fix types of params

+++ core/network-api/src/main/java/ru/yeahub/network_api/models/GetSpecializationsResponse.kt
@@ -4,5 +4,5 @@ data class GetSpecializationsResponse(
     val page: Long?,
     val limit: Long?,
     val data: List<NestedSpecializationResponse>,
-    val total: Int
+    val total: Long
 )
